chore: renamed fringe_list_util.py to util.py
